// src/openapi/models/event-request-log-dto-pagination-response-wrapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { EventRequestLogDto } from './event-request-log-dto';
import { Pagination } from './pagination';

/**
 * 
 * @export
 * @interface EventRequestLogDtoPaginationResponseWrapper
 */
export interface EventRequestLogDtoPaginationResponseWrapper {
    /**
     * 
     * @type {Array<EventRequestLogDto>}
     * @memberof EventRequestLogDtoPaginationResponseWrapper
     */
    data: Array<EventRequestLogDto>;
    /**
     * 
     * @type {Pagination}
     * @memberof EventRequestLogDtoPaginationResponseWrapper
     */
    pagination: Pagination;
}